bugfix for fish version 4
It seems that the update to fish version 4 made the `status current-command` command append a newline at the end of the command. Alternatively the test command could have been made aware of such a newline. I did not investigate this.

Whatever the exact change in behaviour resulted from, I now added a call to `tr` to remove that newline. Thus it now works on fisher version 4.